+++ packages/eas-cli/src/submissions/android/AndroidSubmitter.ts
@@ -14,7 +14,10 @@ import { AndroidSubmissionConfig, ReleaseStatus, ReleaseTrack } from './AndroidS
 import { ServiceAccountSource, getServiceAccountAsync } from './ServiceAccountSource';
 
 export interface AndroidSubmissionOptions
-  extends Pick<AndroidSubmissionConfig, 'track' | 'releaseStatus' | 'projectId'> {
+  extends Pick<
+    AndroidSubmissionConfig,
+    'track' | 'releaseStatus' | 'projectId' | 'changesNotSentForReview'
+  > {
   androidPackageSource: AndroidPackageSource;
   archiveSource: ArchiveSource;
   serviceAccountSource: ServiceAccountSource;
@@ -66,14 +69,14 @@ class AndroidSubmitter extends BaseSubmitter<AndroidSubmissionContext, AndroidSu
     { archive, androidPackage, serviceAccountPath }: ResolvedSourceOptions
   ): Promise<AndroidSubmissionConfig> {
     const serviceAccount = await fs.readFile(serviceAccountPath, 'utf-8');
-    const { track, releaseStatus, projectId } = options;
+    const { track, releaseStatus, projectId, changesNotSentForReview } = options;
 
-    // structuring order affects table rows order
     return {
       androidPackage,
       archiveUrl: archive.location,
       archiveType: archive.type as AndroidArchiveType,
       track,
+      changesNotSentForReview,
       releaseStatus,
       projectId,
       serviceAccount,
@@ -84,12 +87,14 @@ class AndroidSubmitter extends BaseSubmitter<AndroidSubmissionContext, AndroidSu
     options: AndroidSubmissionOptions,
     { archive, androidPackage, serviceAccountPath }: ResolvedSourceOptions
   ): Summary {
-    const { projectId, track, releaseStatus } = options;
+    const { projectId, track, releaseStatus, changesNotSentForReview } = options;
 
+    // structuring order affects table rows order
     return {
       projectId,
       androidPackage,
       track,
+      changesNotSentForReview,
       releaseStatus,
       archiveType: archive.type as AndroidArchiveType,
       serviceAccountPath,
@@ -105,6 +110,7 @@ type Summary = {
   track: ReleaseTrack;
   releaseStatus?: ReleaseStatus;
   projectId?: string;
+  changesNotSentForReview?: boolean;
 } & ArchiveSourceSummaryFields;
 
 const SummaryHumanReadableKeys: Record<keyof Summary, string> = {
@@ -113,6 +119,7 @@ const SummaryHumanReadableKeys: Record<keyof Summary, string> = {
   archiveUrl: 'Download URL',
   archiveType: 'Archive type',
   serviceAccountPath: 'Google Service Key',
+  changesNotSentForReview: 'Changes not sent for a review',
   track: 'Release track',
   releaseStatus: 'Release status',
   projectId: 'Project ID',